requirements.txt
black==22.12.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.3
platformdirs==2.6.2
python-dotenv==0.21.1
SQLAlchemy==1.4.46
tomli==2.0.1
Werkzeug==2.2.2
